apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "users-api"
version = "0.2.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]",
  "gunicorn",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg",
  "psycopg[binary]",
  "alembic",
  "pydantic>=2",
  "python-json-logger",
  "structlog",
  "sentry-sdk[fastapi]>=2.6.0",
  "opentelemetry-sdk>=1.26.0",
  "opentelemetry-exporter-otlp>=1.26.0",
  "opentelemetry-instrumentation-fastapi>=0.47b0",
  "opentelemetry-instrumentation-logging>=0.47b0",
  "httpx>=0.27.0",
  "pytest",
  "pytest-asyncio",
]
[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
